feat(api): finalizando autenticação

+++ controllers/feed.js
@@ -1,15 +1,10 @@
 const fs = require('fs');
 const path = require('path');
+
 const { validationResult } = require('express-validator/check');
 
 const Post = require('../models/post');
-
-const clearImage = (filePath) => {
-  const filePathResult = path.join(__dirname, '..', filePath);
-  fs.unlink(filePathResult, (err) => {
-    console.log(err);
-  });
-};
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -31,10 +26,10 @@ exports.getPosts = (req, res, next) => {
       });
     })
     .catch((err) => {
-      const errorToSend = err;
-      if (!errorToSend.statusCode) errorToSend.statusCode = 500;
-
-      next(errorToSend);
+      if (!err.statusCode) {
+        err.statusCode = 500;
+      }
+      next(err);
     });
 };
 
@@ -52,26 +47,33 @@ exports.createPost = (req, res, next) => {
   }
   const imageUrl = req.file.path;
   const { title, content } = req.body;
+  let creator;
   const post = new Post({
     title,
     content,
     imageUrl,
-    creator: { name: 'YSOUZAS' },
+    creator: req.userId,
   });
   post
     .save()
+    .then((result) => User.findById(req.userId))
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
     .then((result) => {
       res.status(201).json({
         message: 'Post created successfully!',
-        post: result,
+        post,
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
-      const errorToSend = err;
-      if (!errorToSend.statusCode) {
-        errorToSend.statusCode = 500;
+      if (!err.statusCode) {
+        err.statusCode = 500;
       }
-      next(errorToSend);
+      next(err);
     });
 };
 
@@ -87,11 +89,10 @@ exports.getPost = (req, res, next) => {
       res.status(200).json({ message: 'Post fetched.', post });
     })
     .catch((err) => {
-      const errorToSend = err;
-      if (!errorToSend.statusCode) {
-        errorToSend.statusCode = 500;
+      if (!err.statusCode) {
+        err.statusCode = 500;
       }
-      next(errorToSend);
+      next(err);
     });
 };
 
@@ -115,39 +116,36 @@ exports.updatePost = (req, res, next) => {
   }
   Post.findById(postId)
     .then((post) => {
-      const postToUpdate = post;
-      if (!postToUpdate) {
+      if (!post) {
         const error = new Error('Could not find post.');
         error.statusCode = 404;
         throw error;
       }
-      if (imageUrl !== postToUpdate.imageUrl) {
-        clearImage(postToUpdate.imageUrl);
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
       }
-      postToUpdate.title = title;
-      postToUpdate.imageUrl = imageUrl;
-      postToUpdate.content = content;
-      return postToUpdate.save();
+      if (imageUrl !== post.imageUrl) {
+        clearImage(post.imageUrl);
+      }
+      post.title = title;
+      post.imageUrl = imageUrl;
+      post.content = content;
+      return post.save();
     })
     .then((result) => {
       res.status(200).json({ message: 'Post updated!', post: result });
     })
     .catch((err) => {
-      const errorToSend = err;
-      if (!errorToSend.statusCode) {
-        errorToSend.statusCode = 500;
+      if (!err.statusCode) {
+        err.statusCode = 500;
       }
-      next(errorToSend);
+      next(err);
     });
 };
 
 exports.deletePost = (req, res, next) => {
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    const error = new Error('Validation failed, entered data is incorrect.');
-    error.statusCode = 422;
-    throw error;
-  }
   const { postId } = req.params;
   Post.findById(postId)
     .then((post) => {
@@ -156,18 +154,32 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
       // Check logged in user
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
+    .then((result) => User.findById(req.userId))
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
+    })
     .then((result) => {
-      console.log(result);
       res.status(200).json({ message: 'Deleted post.' });
     })
     .catch((err) => {
-      const errorToSend = err;
-      if (!errorToSend.statusCode) errorToSend.statusCode = 500;
-      errorToSend.message = err.message;
-      next(errorToSend);
+      if (!err.statusCode) {
+        err.statusCode = 500;
+      }
+      next(err);
     });
+};
+
+const clearImage = (filePath) => {
+  filePath = path.join(__dirname, '..', filePath);
+  fs.unlink(filePath, (err) => console.log(err));
 };
